Reverted to .cpp files
