+++ editor/meson.build
@@ -7,11 +7,11 @@ endif
 executable(
     'wren_editor',
     'src/main.cpp',
-    # 'src/editor.cxx',
+    'src/editor.cpp',
     'src/ui.cpp',
     'src/scene_panel.cpp',
-    # 'src/inspector_panel.cxx',
-    # 'src/filesystem_panel.cxx',
+    'src/inspector_panel.cpp',
+    'src/filesystem_panel.cpp',
     dependencies: deps,
     install: true,
