+++ imgui.ini
@@ -34,8 +34,8 @@ Collapsed=0
 DockId=0x00000005,0
 
 [Window][Inspector]
-Pos=2058,27
-Size=160,1360
+Pos=1845,27
+Size=373,1360
 Collapsed=0
 DockId=0x00000002,0
 
@@ -47,7 +47,7 @@ DockId=0x00000006,0
 
 [Window][Viewer]
 Pos=129,27
-Size=1927,1360
+Size=1714,1360
 Collapsed=0
 DockId=0x00000004,0
 
@@ -160,10 +160,10 @@ Column 13 Width=19
 
 [Docking][Data]
 DockSpace       ID=0xBB06786C Window=0xCB5DF48C Pos=8,27 Size=2210,1360 Split=X
-  DockNode      ID=0x00000001 Parent=0xBB06786C SizeRef=478,1395 Split=X
+  DockNode      ID=0x00000001 Parent=0xBB06786C SizeRef=1835,1395 Split=X
     DockNode    ID=0x00000003 Parent=0x00000001 SizeRef=119,1395 Split=Y
       DockNode  ID=0x00000005 Parent=0x00000003 SizeRef=119,1044 Selected=0xE192E354
       DockNode  ID=0x00000006 Parent=0x00000003 SizeRef=119,349 Selected=0x07E2D2CA
-    DockNode    ID=0x00000004 Parent=0x00000001 SizeRef=357,1395 CentralNode=1 Selected=0x652D1CFE
-  DockNode      ID=0x00000002 Parent=0xBB06786C SizeRef=160,1395 Selected=0xE7039252
+    DockNode    ID=0x00000004 Parent=0x00000001 SizeRef=1714,1395 CentralNode=1 Selected=0x652D1CFE
+  DockNode      ID=0x00000002 Parent=0xBB06786C SizeRef=373,1395 Selected=0xE7039252
